+++ python/mujoco/introspect/enums.py
@@ -609,16 +609,6 @@ ENUMS: Mapping[str, EnumDecl] = dict([
              ('mjS_AWAKE', 1),
          ]),
      )),
-    ('mjtTaskStatus',
-     EnumDecl(
-         name='mjtTaskStatus',
-         declname='enum mjtTaskStatus_',
-         values=dict([
-             ('mjTASK_NEW', 0),
-             ('mjTASK_QUEUED', 1),
-             ('mjTASK_COMPLETED', 2),
-         ]),
-     )),
     ('mjtGeomInertia',
      EnumDecl(
          name='mjtGeomInertia',